Add load options for PRG and CHR banks (#28)
The primary feature of this commit is to add load options for PRG and
CHR banks.  Selecting the Options button in the load dialog will
show (if appropriate for the mapper) choices for the base address
of each PRG bank, and a choice of how to handle CHR banks.

There is a fair bit of refactoring going here to support this, as well
as preparing for more generalized mapper support in the future.  In
particular, there has been some shuffling of code between the NesRom,
GhidraNesLoader, and NesMapper classes to better align them with their
logical responsibilities.  The separate classes for UxROM and MMC4
have been rendered moot as a result, so they have been removed.

I've also started adding more informational messages that will appear
in the log window when loading a ROM.